requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.1.0
python-jose[cryptography]>=3.3.0,<4.0.0
bcrypt==4.0.1
passlib==1.7.4
pytesseract>=0.3.8,<0.4.0
pdf2image>=1.16.0
Pillow>=8.3.2,<9.0.0
spacy>=3.1.0,<4.0.0
redis>=4.0.0,<5.0.0
sqlalchemy>=1.4.23,<2.0.0
psycopg2-binary>=2.9.1,<3.0.0
python-dotenv>=0.19.0,<0.20.0
boto3>=1.18.0,<2.0.0
celery>=5.1.2,<6.0.0
alembic>=1.7.0,<2.0.0

# Test dependencies
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.18.0,<0.19.0
pytest-cov>=3.0.0,<4.0.0
httpx>=0.23.0,<0.24.0
aiofiles>=0.8.0,<0.9.0
pytest-mock>=3.6.1,<4.0.0
PyJWT>=2.3.0,<3.0.0
